requirements.txt
requests
Pillow
filetype
lxml
json5
